% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_timeseries.R
\name{add_timeseries}
\alias{add_timeseries}
\title{Adds observation timeseries to a 'HS' object}
\usage{
add_timeseries(zones, timeseries, name, unit = NULL, IDs = NULL)
}
\arguments{
\item{zones}{An \code{zones} object.}

\item{timeseries}{a data.frame with observations. Must contain column 
\code{Date}.}

\item{name}{Name of the variable}

\item{unit}{The unit of values in \code{timeseries}.}

\item{IDs}{A vector of riverID of the river segments of the columns
in timeseries. If not provided, IDs will be taken from the column names in
\code{timeseries}.}
}
\value{
Returns the \code{HS} object with added list column 
  \code{observation_ts} containing a timeseries, and column 
  \code{observation_station} containing the name of observation stations.
}
\description{
Adds observation timeseries to a HS object. This is needed in order to 
evaluate performance of downscaled timeseries, or to perform data assimilation
combining several downscaled times.
}
