% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildings.R
\docType{data}
\name{buildings}
\alias{buildings}
\title{This is data of the residential building footprints of Münster}
\format{
A data frame with columns:
\describe{
\item{gid}{ID of building.}
\item{funktion}{Type of the building.}
\item{geometry}{Geometry of the building.}
}
}
\source{
\url{https://opendata.stadt-muenster.de/dataset/geb\%C3\%A4udeumrisse}
}
\usage{
buildings
}
\description{
This is data of the residential building footprints of Münster
}
\keyword{datasets}
