% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_source.R
\name{combine_zones}
\alias{combine_zones}
\title{Combine \code{zones} objects or add new runoff input from raster.}
\usage{
combine_zones(
  zones,
  from = NULL,
  rasters = NULL,
  unit = NULL,
  date = NULL,
  timestep = NULL,
  aoi = NULL,
  names = NULL
)
}
\arguments{
\item{zones}{an existing \code{zones} object.}

\item{from}{A \code{zones} object to add from. Optional.}

\item{rasters}{A raster object, or a list of raster objects, or a list of
URIs to files readable by \code{raster} package.}

\item{unit}{Unit of the timeseries, coercable with \code{units::set_units()}.}

\item{date}{Date of the first layer in \code{raster}, or a vector of dates 
with the length of layers in \code{raster}, or a list of vectors of dates,
corresponding to the dates in each raster layer.}

\item{timestep}{Length of timestep in raster. Currently supporting 
\code{"hour"}, \code{"day"}, \code{"month"}. Only needed if \code{date} is
not a vector of dates.}

\item{aoi}{Area of interest as an 'sf' polygon object, 
'SpatialPolygons', or 'SpatialPolygonsDataFrame'. Optional, but 
recommended for all applications.}

\item{names}{A name (names) for the runoff timeserie(s).}
}
\value{
Returns the input \code{zones} object with added columns in 
  runoff_ts column.
}
\description{
Combines the \code{runoff_ts} columns from two zones objects, or adds new
runoff timeseries from a raster. **All** geometries in zones must be present
in the object to be added, or combination will not done.
}
