% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corine_18.R
\docType{data}
\name{corine_18}
\alias{corine_18}
\title{This is data of the land cover of Münster}
\format{
A data frame with columns:
\describe{
\item{CLC18}{ID of land cover type.}
\item{Shape_Leng}{Length of the polygon.}
\item{Shape_Area}{Area of the polygon.}
\item{geometry}{Geometry of the polygon.}
}
}
\source{
\url{https://daten.gdz.bkg.bund.de/produkte/dlm/clc5_2018/aktuell/clc5_2018.utm32s.shape.zip}
}
\usage{
corine_18
}
\description{
This is data of the land cover of Münster
}
\keyword{datasets}
