% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_source.R
\name{create_target}
\alias{create_target}
\title{Constructs a \code{zones} object from given input}
\usage{
create_target(zones, ID)
}
\arguments{
\item{zones}{An \code{sf} object containing polygon representation of the 
spatial units.}

\item{ID}{Column name with unique IDs.}
}
\value{
Returns a data.frame with an added class \code{zones}, and column
  \code{targetID}
}
\description{
Creates a \code{zones} object from sf polygon representing the spatial units 
of data.
}
