% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasymetric.R
\name{dasymetric}
\alias{dasymetric}
\title{dasymetric: a package for advanced areal interpolation}
\description{
Provides functions for area weighed interpolation, dasymetric mapping, 
pycnophylactic interpolation, and combined pycnophylactic-dasymetric
interpolation. Dasymetric modelling is also supported.
}
