% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasymetric_map.R
\name{dasymetric_map}
\alias{dasymetric_map}
\title{Create a dasymetric map}
\usage{
dasymetric_map(x, ancillary_data, target_geom)
}
\arguments{
\item{x}{numeric}

\item{ancillary_data}{sf}

\item{target_geom}{sf}
}
\value{

}
\description{
Create a dasymetric map
}
\examples{
dasymetric_map(a,b,c)
}
