% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasymetric_map_adv.R
\name{dasymetric_map_adv}
\alias{dasymetric_map_adv}
\title{Create a dasymetric map}
\source{
\url{https://github.com/slu-openGIS/areal/pull/27/commits/d86490f6544af4235bdbdf5f51a9cab000d2b78e},
\url{https://www.eea.europa.eu/data-and-maps/data/population-density/mapping-population-density/mapping-population-density/download}
}
\usage{
dasymetric_map_adv(
  target,
  source,
  ancillary_data,
  tid = NULL,
  extensive = NULL
)
}
\arguments{
\item{target}{sf object containing geometry of the desired spatial zones}

\item{source}{sf object that we want to interpolate}

\item{ancillary_data}{sf object containing geometry that helps to better interpolate (i.e. land use, building footprints)}

\item{tid}{Optional string denoting column with unique identifier for \code{target} geometries. Optional, and will otherwise be automatically generated}

\item{extensive}{Required atomic vector of strings denoting columns in \code{source} with extensive variables (i.e. count data)}
}
\value{
sf
}
\description{
IMMATURE! RESULTS IN WAY TOO LESS POPULATION! Advanced dasymetric mapping that uses CORINE land use classes to make more valuable predictions on population counts. R
}
\examples{
source_geom = sf::st_union(population_counts)
source=sf::st_sf(ID=1,population=sum(population_counts["population"]$population),source_geom)
dasymetric_map_adv(population_counts, source, corine_18, extensive = "population")
}
\references{
Peedell, Steve (1999): Mapping Population Density. Distribution of Population using CORINE Land Cover. Ispra.
}
