% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/districts.R
\docType{data}
\name{districts}
\alias{districts}
\title{This is data of the districts of Münster}
\format{
A data frame with columns:
\describe{
\item{OBJECTID}{Object ID.}
\item{STATIST_BE}{Number and Name of the district.}
\item{NR_STATIST}{Number of the district.}
\item{NAME_STATI}{Name of the district.}
\item{STADTBEZIR}{Number and name of the borough.}
\item{STAND}{timestamp.}
\item{SHAPE_AREA}{Area of the district.}
\item{SHAPE_LEN}{Length of the district.}
\item{geometry}{Geometry of the district.}
}
}
\source{
\url{https://opendata.stadt-muenster.de/dataset/geometriedaten-der-kleinr\%C3\%A4umigen-gebietsgliederung/resource/cf908ea8-1f2c-4e6e-b4dc}, downloaded 2021-08-12
}
\usage{
districts
}
\description{
This is data of the districts of Münster
}
\keyword{datasets}
