% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_counts.R
\docType{data}
\name{population_counts}
\alias{population_counts}
\title{This is data of the population of Münster}
\format{
A data frame with columns:
\describe{
\item{NR_STATIST}{Number of the district.}
\item{NAME_STATI}{Name of the district.}
\item{STADTBEZIR}{Number and name of the borough.}
\item{pop_counts}{Population number.}
\item{SHAPE_AREA}{Area of the district.}
\item{geometry}{Geometry of the district.}
}
}
\source{
\url{https://opendata.stadt-muenster.de/dataset/geometriedaten-der-kleinr\%C3\%A4umigen-gebietsgliederung/resource/cf908ea8-1f2c-4e6e-b4dc}, downloaded 2021-08-12
}
\usage{
population_counts
}
\description{
This is data of the population of Münster
}
\keyword{datasets}
