% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_source.R
\name{combine_zones}
\alias{combine_zones}
\title{Combine \code{zones} objects or add new runoff input from raster.}
\usage{
combine_zones(
  zones,
  from = NULL,
  rasters = NULL,
  unit = NULL,
  date = NULL,
  timestep = NULL,
  aoi = NULL,
  names = NULL
)
}
\arguments{
\item{zones}{an existing \code{zones} object.}

\item{from}{A \code{zones} object to add from. Optional.}
}
\value{
Returns the input \code{zones} object with added columns in 
  runoff_ts column.
}
\description{
Combines the \code{runoff_ts} columns from two zones objects, or adds new
runoff timeseries from a raster. **All** geometries in zones must be present
in the object to be added, or combination will not done.
}
