% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasymetric_map.R
\name{dasymetric_map}
\alias{dasymetric_map}
\title{Create a dasymetric map}
\usage{
dasymetric_map(target, source, ancillary_data, tid = NULL, extensive = NULL)
}
\arguments{
\item{target}{sf object containing geometry of the desired spatial zones}

\item{source}{sf object that we want to interpolate}

\item{ancillary_data}{sf object containing geometry that helps to better interpolate (i.e. land use, building footprints)}

\item{tid}{Optional string denoting column with unique identifier for \code{target} geometries. Optional, and will otherwise be automatically generated}

\item{extensive}{Required atomic vector of strings denoting columns in \code{source} with extensive variables (i.e. count data)}
}
\value{
sf
}
\description{
Binary dasymetric mapping that uses urban areas to make more valuable predictions on population counts
}
\examples{
source_geom = sf::st_union(population_counts)
source=sf::st_sf(ID=1,population=sum(population_counts["population"]$population),source_geom)
urban = prep_landuse(corine_18)
dasymetric_map(population_counts, source, urban, extensive = "population")
}
