% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_landuse.R
\name{prep_landuse}
\alias{prep_landuse}
\title{Prepare land use data for binary dasymetric mapping}
\usage{
prep_landuse(x, class_ids = c(111, 112))
}
\arguments{
\item{x}{sf containing \emph{CORINE} land use data}

\item{class_ids}{numeric identifiers that describes the urban CLC-classes (111 is Continuous urban fabric and 112 is Discontinuous urban fabric)}
}
\value{
sf containing geometries where people probably live
}
\description{
Extract features of a \emph{CORINE} land use sf that cover urban area.
}
\examples{
prep_landuse(corine_18)
}
\seealso{
\code{\link{dasymetric_map}} for using the prepared dataset
}
